Create Version 0.4
This is so far the final part of the code, I fixed the list that goes down 1 through 4, to now where it goes through all possible options. This fixed the problem that I was facing. Adding on, I changed the input which stated "Enter your number in radians" to just "Enter your number" I could not figure out how to use pi in the function.